requirements.txt
numpy==1.26.4
PyYAML==6.0.1
sumolib==1.20.0
traci==1.20.0
